# Improve performance of Enum*Windows() interop functions