--- a/src/System.Windows.Forms.Primitives/src/Interop/User32/Interop.EnumThreadWindows.cs
+++ b/src/System.Windows.Forms.Primitives/src/Interop/User32/Interop.EnumThreadWindows.cs
@@ -9,16 +9,34 @@ internal static partial class Interop
 {
     internal static partial class User32
     {
-        public delegate bool EnumThreadWindowsCallback(IntPtr hWnd, IntPtr lParam);
+        public delegate BOOL EnumThreadWindowsCallback(IntPtr hWnd);
+
+        private delegate BOOL EnumThreadWindowsNativeCallback(IntPtr hWnd, IntPtr lParam);
+
+        private static readonly EnumThreadWindowsNativeCallback s_enumThreadWindowsNativeCallback = HandleEnumThreadWindowsNativeCallback;
 
         [DllImport(Libraries.User32, ExactSpelling = true)]
-        public static extern BOOL EnumThreadWindows(uint dwThreadId, EnumThreadWindowsCallback lpfn, IntPtr lParam);
+        private static extern BOOL EnumThreadWindows(uint dwThreadId, EnumThreadWindowsNativeCallback lpfn, IntPtr lParam);
 
-        public static BOOL EnumThreadWindows(uint dwThreadId, EnumThreadWindowsCallback lpfn, IHandle lParam)
+        public static BOOL EnumThreadWindows(uint dwThreadId, EnumThreadWindowsCallback lpfn)
         {
-            BOOL result = EnumThreadWindows(dwThreadId, lpfn, lParam.Handle);
-            GC.KeepAlive(lParam);
-            return result;
+            // We pass a static delegate to the native function and supply the callback as
+            // reference data, so that the CLR doesn't need to generate a native code block for
+            // each callback delegate instance (for storing the closure pointer).
+            var gcHandle = GCHandle.Alloc(lpfn);
+            try
+            {
+                return EnumThreadWindows(dwThreadId, s_enumThreadWindowsNativeCallback, GCHandle.ToIntPtr(gcHandle));
+            }
+            finally
+            {
+                gcHandle.Free();
+            }
+        }
+
+        private static BOOL HandleEnumThreadWindowsNativeCallback(IntPtr hWnd, IntPtr lParam)
+        {
+            return ((EnumThreadWindowsCallback)GCHandle.FromIntPtr(lParam).Target)(hWnd);
         }
     }
 }