--- a/src/System.Windows.Forms/src/System/Windows/Forms/Application.ThreadWindows.cs
+++ b/src/System.Windows.Forms/src/System/Windows/Forms/Application.ThreadWindows.cs
@@ -30,15 +30,12 @@ namespace System.Windows.Forms
             {
                 _windows = new IntPtr[16];
                 _onlyWinForms = onlyWinForms;
-                var callback = new User32.EnumThreadWindowsCallback(Callback);
                 User32.EnumThreadWindows(
                     Kernel32.GetCurrentThreadId(),
-                    callback,
-                    IntPtr.Zero);
-                GC.KeepAlive(callback);
+                    Callback);
             }
 
-            private bool Callback(IntPtr hWnd, IntPtr lparam)
+            private BOOL Callback(IntPtr hWnd)
             {
                 // We only do visible and enabled windows.  Also, we only do top level windows.
                 // Finally, we only include windows that are DNA windows, since other MSO components
@@ -68,7 +65,7 @@ namespace System.Windows.Forms
                     }
                 }
 
-                return true;
+                return BOOL.TRUE;
             }
 
             // Disposes all top-level Controls on this thread