--- a/src/System.Windows.Forms/src/System/Windows/Forms/Application.cs
+++ b/src/System.Windows.Forms/src/System/Windows/Forms/Application.cs
@@ -605,10 +605,7 @@ namespace System.Windows.Forms
 
                     // 248887 we need to send a WM_THEMECHANGED to the top level windows of this application.
                     // We do it this way to ensure that we get all top level windows -- whether we created them or not.
-                    User32.EnumWindowsCallback callback = SendThemeChanged;
-                    User32.EnumWindows(callback, IntPtr.Zero);
-
-                    GC.KeepAlive(callback);
+                    User32.EnumWindows(SendThemeChanged);
                 }
             }
         }
@@ -616,20 +613,20 @@ namespace System.Windows.Forms
         /// <summary>
         ///  This helper broadcasts out a WM_THEMECHANGED to appropriate top level windows of this app.
         /// </summary>
-        private unsafe static bool SendThemeChanged(IntPtr handle, IntPtr extraParameter)
+        private unsafe static BOOL SendThemeChanged(IntPtr handle)
         {
             uint thisPID = Kernel32.GetCurrentProcessId();
             User32.GetWindowThreadProcessId(handle, out uint processId);
             if (processId == thisPID && User32.IsWindowVisible(handle).IsTrue())
             {
-                SendThemeChangedRecursive(handle, IntPtr.Zero);
+                SendThemeChangedRecursive(handle);
                 User32.RedrawWindow(
                     handle,
                     null,
                     IntPtr.Zero,
                     User32.RDW.INVALIDATE | User32.RDW.FRAME | User32.RDW.ERASE | User32.RDW.ALLCHILDREN);
             }
-            return true;
+            return BOOL.TRUE;
         }
 
         /// <summary>
@@ -637,17 +634,15 @@ namespace System.Windows.Forms
         ///  It is assumed at this point that the handle belongs to the current process
         ///  and has a visible top level window.
         /// </summary>
-        private static bool SendThemeChangedRecursive(IntPtr handle, IntPtr lparam)
+        private static BOOL SendThemeChangedRecursive(IntPtr handle)
         {
             // First send to all children...
-            UnsafeNativeMethods.EnumChildWindows(new HandleRef(null, handle),
-                new NativeMethods.EnumChildrenCallback(Application.SendThemeChangedRecursive),
-                NativeMethods.NullHandleRef);
+            User32.EnumChildWindows(handle, Application.SendThemeChangedRecursive);
 
             // Then do myself.
             UnsafeNativeMethods.SendMessage(new HandleRef(null, handle), Interop.WindowMessages.WM_THEMECHANGED, 0, 0);
 
-            return true;
+            return BOOL.TRUE;
         }
 
         /// <summary>