--- a/src/System.Windows.Forms/src/System/Windows/Forms/ComboBox.cs
+++ b/src/System.Windows.Forms/src/System/Windows/Forms/ComboBox.cs
@@ -6078,7 +6078,7 @@ namespace System.Windows.Forms
                 Debug.Assert(!ACWindows.ContainsKey(acHandle));
                 AssignHandle(acHandle);
                 ACWindows.Add(acHandle, this);
-                EnumChildWindows(new HandleRef(this, acHandle), 
+                EnumChildWindows(new HandleRef(this, acHandle),
                     ACNativeWindow.RegisterACWindowRecursive);
             }