## Proposed changes

In `EnumWindows`, `EnumChildWindows`, and `EnumThreadWindows`, instead of passing the supplied callback delegate to the native function, use a (fixed) static delegate, and specify a handle pointer of the supplied callback in the `lParam` parameter (that is available for storing application-provided data).

That way, the CLR no longer needs to allocate memory for each delegate instance in order to store a native code block (that contains the closure pointer), thus improving performance in many cases (see comparison below). (On the other hand, the invoked callback delegate now has a little bit more work to do, by dereferencing a `GCHandle` and calling another method.)

This can especially improve speed when running x86 applications on **ARM64** systems, as it avoids the need for the emulation layer to generate new ARM64 code from the x86 code every time the method is invoked with a new delegate instance.

Note: In `Form.RecreateHandleCore` there was a call to `EnumThreadWindows` where the form handle was passed as `lParam` data, but I think it makes more sense to store it in the callback closure since a new instance of the delegate is created anyway.

Note: I'm not sure about the naming/placement of the delegate types. In the native declarations, all functions use `WNDENUMPROC` as callback type.

<!-- We are in TELL-MODE the following section must be completed -->

## Customer Impact

- Better performance and reduced memory footprint in many cases when using code that calls `EnumWindows`, `EnumChildWindows` or `EnumThreadWindows`.
In specific cases (when the callback is called very often), performance might be slower due to the additional work the callback has to do.

## Regression? 

- No

## Risk

-

<!-- end TELL-MODE -->


## Test methodology <!-- How did you ensure quality? -->

### Performance
- Manual testing using a code like this (for `EnumChildWindows`):
```c#
    class Program
    {
        [STAThread]
        private static void Main()
        {
            var f = new Form();
            f.Controls.Add(new Button());
            f.Controls.Add(new CheckBox());
            f.Controls.Add(new RadioButton());

            var handle = f.Handle;

            f.Show();

            var sw = new Stopwatch();
            for (int i = -5; i < 100000; i++)
            {
                if (i == 0)
                    sw.Start();

                Interop.User32.EnumChildWindows(
                    handle, new EnumChildWindowsTest().HandleCallback);
            }
            sw.Stop();
            f.Close();
            f.Dispose();
            Console.WriteLine("Elapsed: " + sw.ElapsedMilliseconds);            
        }

        private class EnumChildWindowsTest
        {
            private int count;

            public Interop.BOOL HandleCallback(IntPtr hWnd)
            {
                count++;
                return Interop.BOOL.TRUE;
            }
        }
    }
```

**Average time** before/after this PR:
- **x86** app on a **x64** system (Windows 10 Version 1909)
  - Before: ~660 ms
  - After: ~370 ms
  - Factor: 1.8x
- **x64** app:
  - Before: ~400 ms
  - After: ~280 ms
  - Factor: 1.4x
- **x86** app on a (slow) **ARM64** machine (Windows 10 on ARM on Raspberry Pi 3 B+):
  - Before: ~27.0s
  - After: ~5.4s
  - Factor: **5.0x**

Note: When the callback is called more often (e.g. about 200 times when calling `EnumWindows`), the timing improvement vanishes with a x64 app, probably because the callback delegate has now a bit more work to do.


## Test environment(s) <!-- Remove any that don't apply -->

.NET Core SDK (reflecting any global.json):
 Version:   5.0.100-alpha1-016073
 Commit:    bd7bf8c6d2

Runtime Environment:
 OS Name:     Windows
 OS Version:  10.0.18363
 OS Platform: Windows
 RID:         win10-x64
 Base Path:   C:\Program Files\dotnet\sdk\5.0.100-alpha1-016073\

Host (useful for support):
  Version: 5.0.0-alpha.1.19564.1
  Commit:  c77948d92a

.NET Core SDKs installed:
  3.1.100 [C:\Program Files\dotnet\sdk]
  5.0.100-alpha1-016073 [C:\Program Files\dotnet\sdk]

###### Microsoft Reviewers: [Open in CodeFlow](http://wpcp.azurewebsites.net/CodeFlowProtocolProxy2.php?pullrequest=https://github.com/dotnet/winforms/pull/2639)